backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
motor==3.3.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
certifi==2024.2.2
dnspython==2.6.1
scikit-learn==1.8.0
joblib==1.5.3
pandas==2.3.1
numpy==2.2.6
httpx==0.28.1
